+++ src/okfn_iati/activities/process_xml/__init__.py
@@ -1,0 +1,1 @@
+"""IATI XML to CSV extractors module __init__."""
