+++ src/okfn_iati/organisations/__init__.py
@@ -1,0 +1,33 @@
+"""
+IATI Organisation Processing Module.
+
+Provides classes for generating and converting IATI organisation XML and CSV files.
+
+Main classes:
+    - IatiOrganisationXMLGenerator: Generate IATI organisation XML from records
+    - IatiOrganisationMultiCsvConverter: Convert between multi-CSV and organisation XML
+    - OrganisationRecord: Data model for organisation information
+    - OrganisationBudget: Data model for organisation budget
+    - OrganisationExpenditure: Data model for organisation expenditure
+    - OrganisationDocument: Data model for organisation document
+"""
+
+from .base import (
+    IatiOrganisationXMLGenerator,
+    IatiOrganisationCSVConverter,
+    IatiOrganisationMultiCsvConverter,
+    OrganisationRecord,
+    OrganisationBudget,
+    OrganisationExpenditure,
+    OrganisationDocument
+)
+
+__all__ = [
+    "IatiOrganisationXMLGenerator",
+    "IatiOrganisationCSVConverter",
+    "IatiOrganisationMultiCsvConverter",
+    "OrganisationRecord",
+    "OrganisationBudget",
+    "OrganisationExpenditure",
+    "OrganisationDocument"
+]
