+++ src/okfn_iati/organisations/process_xml/__init__.py
@@ -1,0 +1,1 @@
+"""IATI organisation XML extraction module."""
